requirements.txt
CacheControl==0.12.11
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cmake==3.22.4
colorama==0.4.4
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
firebase-admin==5.2.0
Flask==2.1.2
Flask-Cors==3.0.10
google-api-core==2.8.1
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.5.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.3
numpy==1.21.6
opencv-python==4.5.5.64
Pillow==9.1.1
proto-plus==1.20.5
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.27.1
rsa==4.8
six==1.16.0
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
